fix(p3): marker liveness precedence + concurrent-revert guards + last coverage legs (codex P2 pair + lint)
- fsutil: positive owner liveness makes a marker busy regardless of age
  (alive probe is never overridden by the 2-minute rule; age only decides
  when liveness cannot be established); W20 start-time reuse detection and
  atomic takeover preserved
- history: per-operation rev-op:<id> serialization around revertFile plus
  fresh-row revalidation — consumed entries skip cleanly instead of failing
  the second concurrent revert mid-loop, leaving journal and op status
  coherent; op serializer, dest locks, busy marker claim order documented
- lint: dropped the unused lstatRestoreSource result and the redundant
  uint64 conversion
- coverage: canonicalize-failure fallback, cleanup-walk Remove failure and
  root-boundary legs covered end of tail
Strict patch gate verified 100% (1778/0/0) on the full-suite profile;
golangci-lint clean.

+++ internal/history/restore_source_identity_unix.go
@@ -15,5 +15,5 @@ func restoreSourceIdentity(info os.FileInfo) (device, inode uint64, ok bool) {
 	if !ok || stat == nil {
 		return 0, 0, false
 	}
-	return uint64(stat.Dev), uint64(stat.Ino), true
+	return uint64(stat.Dev), stat.Ino, true
 }
